move tests out of lib.rs and into tests/test.rs
